requirements.txt
livekit==1.0.23
livekit-agents[silero,openai,cartesia,deepgram]==1.3.11
livekit-api==1.1.0
livekit-protocol==1.1.1
livekit-plugins-openai==1.3.11
livekit-plugins-cartesia==1.3.11
livekit-plugins-deepgram==1.3.11
livekit-plugins-silero==1.3.11
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-turn-detector==1.3.11
livekit-plugins-azure==1.3.11
openai==1.109.1
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-oauthlib==1.3.0
google-auth-httplib2==0.3.0
supabase==2.27.0
python-dotenv==1.0.1
phonenumbers==9.0.15
email-validator==2.3.0
python-dateutil==2.9.0.post0
tzdata==2025.2
aiohttp==3.12.15
fastapi==0.111.0
uvicorn==0.29.0
asyncpg==0.30.0
paramiko==3.5.1
tenacity==9.1.2
twilio==9.8.3
python-multipart==0.0.20
sentry-sdk[asyncio]==2.39.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-mock==3.15.1
